Implement WS API workflow for checking quality gate.
Since SQ 5.3, a "report-task.txt" file is saved in the working directory
with the CE task ID.  Query /api/ce/task?id=<ceTaskId> until success to
retrieve the analysis ID, then query
/api/qualitygates/project_status?analysisId=<analysisId> for the quality
gate status.

Because it is possible for the CE task to not be complete on first
query, additional queries will be made until its completion or until the
limit is reached.

The CE task query limit and interval between queries is configurable.

Breaks the build under any of the following conditions:
- CE task status returns 'FAILED' or 'CANCELED'
- CE task API query limit reached
- Quality gate for project analysis is 'ERROR'
- Any unexpected failure (WS timeout, missing report file, etc.)